# [cuda] faster kernelTransformReduceInnermostDimIndex using cub