@ezyang 2019-06-05
Pasting some prior discussion here:

----

@wesolwsk: if you can point me to a workflow using this operator we can check the SM occupancy before and after

----

Incidentally, the macro for num threads per block is set much too high for pyTorch (1024). I previously found that for Caffe2 lowering it from 512 to 128 had a positive effect on perf for use cases with small to medium sized inputs so it may be worth trying for PyTorch

----

@akyrola: I don't have a workflow, this is just random point optimization from a bootcamp task. But below, I run nvprof to get "achieved occupancy", which I think is the same thing?

Here is my benchmark script:

```
import torch
from random import *
import time
import timeit

n = 1000000
I = torch.tensor([[randint(0, 200) for _ in range(3)] for _ in range(n)]).t().cuda()
V = torch.randn(n).cuda()
size = torch.Size([2000, 1000, 1000])

def fn():
    torch.cuda.synchronize();
    S = torch.sparse_coo_tensor(I, V, size)
    torch.cuda.synchronize();

t = timeit.repeat(fn, repeat=100, number=1)
print(min(t))
```

```
nvprof --metrics achieved_occupancy python sparse_bench.py
```

Before:
```
0.05
==2041189== Profiling application: python sparse_bench.py
==2041189== Profiling result:
==2041189== Metric result:
Invocations                               Metric Name                        Metric Description         Min         Max         Avg
Device "Tesla M40 (0)"
    Kernel: void kernelTransformReduceInnermostDimIndex<long, long, MaxValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
        100                        achieved_occupancy                        Achieved Occupancy    0.249997    0.249998    0.249998
    Kernel: void kernelTransformReduceInnermostDimIndex<long, long, MinValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
        100                        achieved_occupancy                        Achieved Occupancy    0.249998    0.249998    0.249998
times:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   50.02%  2.53547s       100  25.355ms  25.243ms  29.036ms  void kernelTransformReduceInnermostDimIndex<long, long, MinValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
                   49.92%  2.53042s       100  25.304ms  25.223ms  28.981ms  void kernelTransformReduceInnermostDimIndex<long, long, MaxValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
```

After:

```
0.03
==2032478== Profiling application: python sparse_bench.py
==2032478== Profiling result:
==2032478== Metric result:
Invocations                               Metric Name                        Metric Description         Min         Max         Avg
Device "Tesla M40 (0)"
    Kernel: void kernelTransformReduceInnermostDimIndex<long, long, MaxValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
        100                        achieved_occupancy                        Achieved Occupancy    0.124988    0.124990    0.124989
    Kernel: void kernelTransformReduceInnermostDimIndex<long, long, MinValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
        100                        achieved_occupancy                        Achieved Occupancy    0.124988    0.124999    0.124989
times:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   49.63%  164.96ms       100  1.6496ms  1.5968ms  2.0850ms  void kernelTransformReduceInnermostDimIndex<long, long, MinValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
                   49.47%  164.45ms       100  1.6445ms  1.5961ms  1.8330ms  void kernelTransformReduceInnermostDimIndex<long, long, MaxValuePair<long, long>>(long*, long*, long*, unsigned int, unsigned int, thrust::pair<long, long>, long)
```

Interestingly the occupancy is lower on the new version, but it runs 15x faster. Occupancy in isolation is not a good metric because very inefficient algorithm can have a high occupancy.
Let me do some additional testing to ensure the correctness, although I think the other tests cover it well.

----

I tried with 128 threads, and the performance is 2x slower on my (admittedly too simple benchmark). I'll keep it in 256 but introduce a define.

----

Reading your post (very impressive result!), indeed 128 would be better for smaller inputs... i guess it would be best to vary the number based on the input size. but that's too much work for this case.

----

@wesolwsk Which input sizes have you tested? Can you try a variety of input sizes (e.g. 100, 10000, 1000000). For the last, scaling the number of blocks with number of rows may work better than limiting number of blocks to 1024.