@ngimel 2019-06-05
I'm generally uneasy about improvements to legacy reductions. Ideally all reduction operators should be ported to TensorIterator reduction that is achieving close to SOL speed on a wide variety of parameters. max/min/max_index/min_index currently don't go through TensorIterator, but they should. 
Having multiple reduction paths, some using cub, some not, is hard to maintain.