@akyrola 2019-06-05
![Screen Shot 2019-06-05 at 9 55 57 AM](https://user-images.githubusercontent.com/1576151/58974838-4bdb4f80-8778-11e9-8c62-bf283a1bc6f2.png)

Indeed @wesolwsk, the speed depends a lot of the shape. Old version is faster if there are many rows. i was doing this in the context of sparse tensors, where the number of rows for the indices is always very small (as it is the number of tensor dimensions). But definitely should develop alternative kernel for the "skinny" tensors.

But given what @ngimel says, perhaps I should drop this? We though have a lot of CUB usage in the caffe2 side.