@wesolwsk 2019-06-05
@akyrola, nice perf comparison. Maybe we can keep both implementations for now but select based on input dimensions (looks like the new one is better for rows of size > 1000. If TensorIterator reductions can handle both cases well then that would be even better.